validation/requirements.txt
torch>=2.0.0
webdataset>=0.2.86
lpips>=0.1.4
tqdm>=4.66.0
numpy>=1.24.0
Pillow>=10.0.0
diffusers>=0.27.0
transformers>=4.38.0
